Former-commit-id: f51277f893a7dc2507bff7c4285dc58398836ff5

+++ backend/drawing-game-service/src/main/java/com/aioi/drawaing/drawinggameservice/drawing/application/dto/RoundInfo.java
@@ -2,5 +2,5 @@ package com.aioi.drawaing.drawinggameservice.drawing.application.dto;
 
 import java.util.List;
 
-public record RoundInfo(List<String> word, List<Integer> drawOrder) {
+public record RoundInfo(List<String> word, List<Long> drawOrder) {
 }
